refactor: #39 코드 리뷰
 - 도메인 분리
 - Transactional 어노테이션 활용
