+++ src/main/java/com/sparta/springbasicsolo/domain/file/repository/JpaFileRepository.java
@@ -1,6 +1,6 @@
-package com.sparta.springbasicsolo.repository;
+package com.sparta.springbasicsolo.domain.file.repository;
 
-import com.sparta.springbasicsolo.repository.entity.Image;
+import com.sparta.springbasicsolo.domain.file.repository.entity.Image;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaFileRepository extends JpaRepository<Image, Long> {
