+++ src/main/java/com/sparta/springbasicsolo/domain/todo/repository/JpaTodoRepository.java
@@ -1,6 +1,6 @@
-package com.sparta.springbasicsolo.repository;
+package com.sparta.springbasicsolo.domain.todo.repository;
 
-import com.sparta.springbasicsolo.repository.entity.Todo;
+import com.sparta.springbasicsolo.domain.todo.repository.entity.Todo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaTodoRepository extends JpaRepository<Todo, Long> {
